Rename: eventLocation -> event_location, eventShortReview -> event_short_review 패키지 이름 변경
