+++ src/main/java/com/otakumap/domain/event/dto/EventResponseDTO.java
@@ -1,6 +1,6 @@
 package com.otakumap.domain.event.dto;
 
-import com.otakumap.domain.eventLocation.dto.EventLocationResponseDTO;
+import com.otakumap.domain.event_location.dto.EventLocationResponseDTO;
 import com.otakumap.domain.image.dto.ImageResponseDTO;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
