+++ src/main/java/com/otakumap/domain/event/entity/Event.java
@@ -3,7 +3,7 @@ package com.otakumap.domain.event.entity;
 import com.otakumap.domain.event.entity.enums.EventStatus;
 import com.otakumap.domain.event.entity.enums.EventType;
 import com.otakumap.domain.event.entity.enums.Genre;
-import com.otakumap.domain.eventLocation.entity.EventLocation;
+import com.otakumap.domain.event_location.entity.EventLocation;
 import com.otakumap.domain.event_like.entity.EventLike;
 import com.otakumap.domain.image.entity.Image;
 import com.otakumap.domain.mapping.EventAnimation;
