+++ src/main/java/com/otakumap/domain/event_short_review/converter/EventShortReviewConverter.java
@@ -1,10 +1,10 @@
-package com.otakumap.domain.eventShortReview.converter;
+package com.otakumap.domain.event_short_review.converter;
 
 import com.otakumap.domain.user.entity.User;
 import com.otakumap.domain.event.entity.Event;
-import com.otakumap.domain.eventShortReview.dto.EventShortReviewRequestDTO;
-import com.otakumap.domain.eventShortReview.dto.EventShortReviewResponseDTO;
-import com.otakumap.domain.eventShortReview.entity.EventShortReview;
+import com.otakumap.domain.event_short_review.dto.EventShortReviewRequestDTO;
+import com.otakumap.domain.event_short_review.dto.EventShortReviewResponseDTO;
+import com.otakumap.domain.event_short_review.entity.EventShortReview;
 import com.otakumap.domain.image.converter.ImageConverter;
 import org.springframework.data.domain.Page;
