+++ src/main/java/com/otakumap/domain/event_short_review/dto/EventShortReviewRequestDTO.java
@@ -1,4 +1,4 @@
-package com.otakumap.domain.eventShortReview.dto;
+package com.otakumap.domain.event_short_review.dto;
 
 import lombok.Getter;
