+++ src/main/java/com/otakumap/domain/event_short_review/repository/EventShortReviewRepository.java
@@ -1,7 +1,7 @@
-package com.otakumap.domain.eventShortReview.repository;
+package com.otakumap.domain.event_short_review.repository;
 
 import com.otakumap.domain.event.entity.Event;
-import com.otakumap.domain.eventShortReview.entity.EventShortReview;
+import com.otakumap.domain.event_short_review.entity.EventShortReview;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.jpa.repository.JpaRepository;
